sidebar_ui: Move initializing functions to module file.
